+++ Runtime/Utils/Builder/IBuild.cs
@@ -1,0 +1,9 @@
+﻿namespace Strangeman.Utils.Builder
+{
+    public interface IBuild<B>
+    {
+        IBuild<B> With<T>(string memberName, T value);
+        T Get<T>(string memberName);
+        B Build();
+    }
+}