+++ Runtime/Utils/Strategy/IValidationStrategy.cs
@@ -1,0 +1,7 @@
+namespace Strangeman.Utils.Strategy
+{
+    public interface IValidationStrategy<S, R>
+    {
+        bool Validate(S source, R rules);
+    }
+}