+++ Runtime/Utils/Visitor/Visitor.cs
@@ -1,0 +1,12 @@
+namespace Strangeman.Utils.Visitor
+{
+    public interface IVisitor
+    {
+        void Visit<T>(T visitable) where T : IVisitable;
+    }
+
+    public interface IVisitable
+    {
+        void Accept(IVisitor visitor);
+    }
+}